+++ src/components/WordView.tsx
